Оптимизировали запрос, добавили синхронизацию по ID

+++ database/migrations/0001_01_01_000015_add_yc_company_daily_stats_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('yc_company_daily_stats', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('company_id'); // ID компании в YClients
+            $table->unsignedBigInteger('company_id'); // ID компании в YClients
             $table->date('date');                  // Дата среза
 
             // Финансовые показатели
